+++ CODEOWNERS
@@ -236,8 +236,8 @@ homeassistant/components/*/simplisafe.py @bachya
 # T
 homeassistant/components/tahoma.py @philklei
 homeassistant/components/*/tahoma.py @philklei
-homeassistant/components/tellduslive.py @molobrakos @fredrike
-homeassistant/components/*/tellduslive.py @molobrakos @fredrike
+homeassistant/components/tellduslive/*.py @fredrike
+homeassistant/components/*/tellduslive.py @fredrike
 homeassistant/components/tesla.py @zabuldon
 homeassistant/components/*/tesla.py @zabuldon
 homeassistant/components/thethingsnetwork.py @fabaff
