+++ homeassistant/components/cover/tellduslive.py
@@ -8,8 +8,9 @@ https://home-assistant.io/components/cover.tellduslive/
 """
 import logging
 
+from homeassistant.components import tellduslive
 from homeassistant.components.cover import CoverDevice
-from homeassistant.components.tellduslive import TelldusLiveEntity
+from homeassistant.components.tellduslive.entry import TelldusLiveEntity
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -19,7 +20,8 @@ def setup_platform(hass, config, add_entities, discovery_info=None):
     if discovery_info is None:
         return
 
-    add_entities(TelldusLiveCover(hass, cover) for cover in discovery_info)
+    client = hass.data[tellduslive.DOMAIN]
+    add_entities(TelldusLiveCover(client, cover) for cover in discovery_info)
 
 
 class TelldusLiveCover(TelldusLiveEntity, CoverDevice):
@@ -33,14 +35,11 @@ class TelldusLiveCover(TelldusLiveEntity, CoverDevice):
     def close_cover(self, **kwargs):
         """Close the cover."""
         self.device.down()
-        self.changed()
 
     def open_cover(self, **kwargs):
         """Open the cover."""
         self.device.up()
-        self.changed()
 
     def stop_cover(self, **kwargs):
         """Stop the cover."""
         self.device.stop()
-        self.changed()
