+++ homeassistant/components/tellduslive/__init__.py
@@ -4,25 +4,22 @@ Support for Telldus Live.
 For more details about this component, please refer to the documentation at
 https://home-assistant.io/components/tellduslive/
 """
-from datetime import datetime, timedelta
+from datetime import timedelta
 import logging
 
 import voluptuous as vol
 
 from homeassistant.components.discovery import SERVICE_TELLDUSLIVE
-from homeassistant.const import (
-    ATTR_BATTERY_LEVEL, CONF_HOST, CONF_TOKEN, DEVICE_DEFAULT_NAME,
-    EVENT_HOMEASSISTANT_START)
+from homeassistant.const import CONF_HOST, CONF_TOKEN
 from homeassistant.helpers import discovery
 import homeassistant.helpers.config_validation as cv
-from homeassistant.helpers.entity import Entity
-from homeassistant.helpers.event import track_point_in_utc_time
-from homeassistant.util.dt import utcnow
+from homeassistant.helpers.dispatcher import dispatcher_send
+from homeassistant.helpers.event import track_time_interval
 from homeassistant.util.json import load_json, save_json
 
-APPLICATION_NAME = 'Home Assistant'
+from .const import DOMAIN, SIGNAL_UPDATE_ENTITY
 
-DOMAIN = 'tellduslive'
+APPLICATION_NAME = 'Home Assistant'
 
 REQUIREMENTS = ['tellduslive==0.10.4']
 
@@ -47,9 +44,6 @@ CONFIG_SCHEMA = vol.Schema({
             vol.All(cv.time_period, vol.Clamp(min=MIN_UPDATE_INTERVAL)))
     }),
 }, extra=vol.ALLOW_EXTRA)
-
-
-ATTR_LAST_UPDATED = 'time_last_updated'
 
 CONFIG_INSTRUCTIONS = """
 To link your TelldusLive account:
@@ -146,7 +140,7 @@ def setup(hass, config, session=None):
         if not supports_local_api(device):
             _LOGGER.debug('Tellstick does not support local API')
             # Configure the cloud service
-            hass.async_add_job(request_configuration)
+            hass.add_job(request_configuration)
             return
 
         _LOGGER.debug('Tellstick does support local API')
@@ -189,18 +183,17 @@ def setup(hass, config, session=None):
         return True
 
     if not session.is_authorized:
-        _LOGGER.error(
-            'Authentication Error')
+        _LOGGER.error('Authentication Error')
         return False
 
     client = TelldusLiveClient(hass, config, session)
-
     hass.data[DOMAIN] = client
+    client.update()
 
-    if session:
-        client.update()
-    else:
-        hass.bus.listen(EVENT_HOMEASSISTANT_START, client.update)
+    interval = config.get(DOMAIN, {}).get(CONF_UPDATE_INTERVAL,
+                                          DEFAULT_UPDATE_INTERVAL)
+    _LOGGER.debug('Update interval %s', interval)
+    track_time_interval(hass, client.update, interval)
 
     return True
 
@@ -210,27 +203,15 @@ class TelldusLiveClient:
 
     def __init__(self, hass, config, session):
         """Initialize the Tellus data object."""
-        self.entities = []
+        self._known_devices = set()
 
         self._hass = hass
         self._config = config
-
-        self._interval = config.get(DOMAIN, {}).get(
-            CONF_UPDATE_INTERVAL, DEFAULT_UPDATE_INTERVAL)
-        _LOGGER.debug('Update interval %s', self._interval)
         self._client = session
 
     def update(self, *args):
-        """Periodically poll the servers for current state."""
-        _LOGGER.debug('Updating')
-        try:
-            self._sync()
-        finally:
-            track_point_in_utc_time(
-                self._hass, self.update, utcnow() + self._interval)
-
-    def _sync(self):
         """Update local list of devices."""
+        _LOGGER.debug('Updating')
         if not self._client.update():
             _LOGGER.warning('Failed request')
 
@@ -254,9 +235,8 @@ class TelldusLiveClient:
             discovery.load_platform(
                 self._hass, component, DOMAIN, [device_id], self._config)
 
-        known_ids = {entity.device_id for entity in self.entities}
         for device in self._client.devices:
-            if device.device_id in known_ids:
+            if device.device_id in self._known_devices:
                 continue
             if device.is_sensor:
                 for item in device.items:
@@ -265,9 +245,9 @@ class TelldusLiveClient:
             else:
                 discover(device.device_id,
                          identify_device(device))
+            self._known_devices.add(device.device_id)
 
-        for entity in self.entities:
-            entity.changed()
+        dispatcher_send(self._hass, SIGNAL_UPDATE_ENTITY)
 
     def device(self, device_id):
         """Return device representation."""
@@ -276,91 +256,3 @@ class TelldusLiveClient:
     def is_available(self, device_id):
         """Return device availability."""
         return device_id in self._client.device_ids
-
-
-class TelldusLiveEntity(Entity):
-    """Base class for all Telldus Live entities."""
-
-    def __init__(self, hass, device_id):
-        """Initialize the entity."""
-        self._id = device_id
-        self._client = hass.data[DOMAIN]
-        self._client.entities.append(self)
-        self._name = self.device.name
-        _LOGGER.debug('Created device %s', self)
-
-    def changed(self):
-        """Return the property of the device might have changed."""
-        if self.device.name:
-            self._name = self.device.name
-        self.schedule_update_ha_state()
-
-    @property
-    def device_id(self):
-        """Return the id of the device."""
-        return self._id
-
-    @property
-    def device(self):
-        """Return the representation of the device."""
-        return self._client.device(self.device_id)
-
-    @property
-    def _state(self):
-        """Return the state of the device."""
-        return self.device.state
-
-    @property
-    def should_poll(self):
-        """Return the polling state."""
-        return False
-
-    @property
-    def assumed_state(self):
-        """Return true if unable to access real state of entity."""
-        return True
-
-    @property
-    def name(self):
-        """Return name of device."""
-        return self._name or DEVICE_DEFAULT_NAME
-
-    @property
-    def available(self):
-        """Return true if device is not offline."""
-        return self._client.is_available(self.device_id)
-
-    @property
-    def device_state_attributes(self):
-        """Return the state attributes."""
-        attrs = {}
-        if self._battery_level:
-            attrs[ATTR_BATTERY_LEVEL] = self._battery_level
-        if self._last_updated:
-            attrs[ATTR_LAST_UPDATED] = self._last_updated
-        return attrs
-
-    @property
-    def _battery_level(self):
-        """Return the battery level of a device."""
-        from tellduslive import (BATTERY_LOW,
-                                 BATTERY_UNKNOWN,
-                                 BATTERY_OK)
-        if self.device.battery == BATTERY_LOW:
-            return 1
-        if self.device.battery == BATTERY_UNKNOWN:
-            return None
-        if self.device.battery == BATTERY_OK:
-            return 100
-        return self.device.battery  # Percentage
-
-    @property
-    def _last_updated(self):
-        """Return the last update of a device."""
-        return str(datetime.fromtimestamp(self.device.lastUpdated)) \
-            if self.device.lastUpdated else None
-
-    @property
-    def unique_id(self) -> str:
-        """Return a unique ID."""
-        return self._id
