+++ homeassistant/components/tellduslive/const.py
@@ -1,0 +1,5 @@
+"""Consts used by TelldusLive."""
+
+DOMAIN = 'tellduslive'
+
+SIGNAL_UPDATE_ENTITY = 'tellduslive_update'
